currently working on the new_job_order.kt need to modify databse schema specifically with the relationships between entities. need to refine activity_new_job_order.xml kay mu lampas ang plus button if mag add og fabcon.
customer search functional
plus and minus button functional

prices for addons not yet fetched
number of loads not being displayed
total weight not being divided into separate load.

+++ app/src/main/java/com/odessy/srlaundry/entities/JobOrder.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class JobOrder(@PrimaryKey(autoGenerate = true) val id: Int = 0,
     val customerName: String,
     val weight: Double,
-    val loads: Double,
+    val loads: Int,
     val addOnDetergent: Int,
     val addOnFabricConditioner: Int,
     val addOnBleach: Int,
